+++ main.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"flag"
 	"log"
-	"terraform-provider-microsoft365wp/workplace"
+	"terraform-provider-microsoft365wp/workplace/provider"
 
 	"github.com/hashicorp/terraform-plugin-framework/providerserver"
 )
@@ -19,7 +19,7 @@ func main() {
 	flag.BoolVar(&debug, "debug", false, "set to true to run the provider with support for debuggers like delve")
 	flag.Parse()
 
-	err := providerserver.Serve(context.Background(), workplace.New, providerserver.ServeOpts{
+	err := providerserver.Serve(context.Background(), provider.New, providerserver.ServeOpts{
 		Address: "registry.terraform.io/terraprovider/microsoft365wp",
 		Debug:   debug,
 	})