+++ workplace/services/generate.go
@@ -1,0 +1,3 @@
+package services
+
+//go:generate sh -c "cd ../../tools/edmx2go/edmx2go-go && go run . -all"